Move authentication/file routes into Blueprints to modularize app
